Allow alternative identifying keywords (#979)
* added test to ensure that an identifying element with a null value (None)
retains that after anonymisation.  This test fails until the commit in the working
code (to follow immediately)

* if an identifying element has a value that is null, or is an empty string, or an
  empty list, don't attempt to modify the value to anonymise it.
  there is no identifying information, and no need to put in some other value

* switched to logging string formatting rather than fstring to avoid linter
complaint

* offer means to provide alternative list of identifying keywords
useful for dynamic additions if the standard changes or keywords were missed
useful for including UIDs either for hardcoded anonymisation or for
pseudonymisation

* docstring entries for replacement_strategy (catch up) and for
alternative_identifying_keywords

* add test of use of alternative list of identifying keywords by appending
SOPInstanceUID to the default list.  Apply hardcode/default anonymisation and
assert that the SOP Instance UID in the anonymised data is not the same as
the original/test data

* refactor per Pull Request comments

* safeguard the default identifying keyword list against accidental mutation
when accessed through the more or less public interface to the list

* safeguard (alternative) identifying keywords against being mutated
unintentionally

* make the output of the privately accessible method for default identifying
keywords immutable

Former-commit-id: e2a6429b3050c7075a831128a8e94a7aa18b947f