Update test_icom_cli.py
Separate the download of the zip files into a separate function to ensure the mock ICOM server is initialised before the 'pymedphys icom listen' command is executed. This prevents the listener from timing out and hanging for 15 minutes before retrying, which could happen if the listener starts before the mock server is ready.